stock_predictor/requirements.txt
yfinance==0.2.28
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
tensorflow==2.15.0
flask==3.0.0
flask-socketio==5.3.6
plotly==5.18.0
dash==2.16.1
dash-bootstrap-components==1.5.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
smtplib-ssl==0.0.1
twilio==8.12.0
websocket-client==1.7.0
threading-timer==0.1.0
